apm: Averaged Prediction Models

In panel data settings, specifies set of candidate models, fits them to data from pre-treatment validation periods, and selects model as average over candidate models, weighting each by posterior probability of being most robust given its differential average prediction errors in pre-treatment validation periods. Subsequent estimation and inference of causal effect's bounds accounts for both model and sampling uncertainty, and calculates the robustness changepoint value at which bounds go from excluding to including 0. Package also includes a range of diagnostic plots, such as those illustrating models' differential average prediction errors and the posterior distribution of which model is most robust.

See also

Author

Maintainer: Noah Greifer ngreifer@iq.harvard.edu (ORCID)

Authors:

Thomas Leavitt thomas.leavitt@baruch.cuny.edu (ORCID)

Laura Hatfield hatfield@hcp.med.harvard.edu (ORCID)