Dataset on Annual Homicide Rates

A dataset of homicide rates across 9 states from 1994 to 2008. Missouri repealed its permit-to-purchase (PTP) law in 2007.

Usage

ptpdata

Format

A dataframe of 7 variables with 135 observations (9 states, 15 years).

Details

details for dataset2

state

the name of the state; states present include Arkansas, Illinois, Iowa, Kansas, Kentucky, Missouri, Nebraska, Oklahoma, and Tennessee.

year

the year of each observation. Years range from 1994 to 2008.

deaths

the number of homicide deaths in the corresponding state in the corresponding year.

crude_rate

the homicide rate in the corresponding state in the corresponding year.

age_adj_rate

the age-adjusted homicide rate in the corresponding state in the corresponding year.

group

whether each observation belongs to the "to-be-treated" group; 1 for Missouri and 0 for all other states.

treat

whether each observation is treated; 1 for Missouri in 2008 and 0 otherwise.